Proof Of Concept
This is my proof of concept for this months project. I am just showing my ability to display 4 songs, artists and albums based off of the users input. I have no styling or structure this is to just show that I can make the api call and do what I plan on doing for this months portfolio project.