Progress
I have gottten the base of my home page functionality down I just need to add overlays so that the song data is displayed on a hover effect. I also have gotten it to upload search content whenever an item is searched. I still have to add the album search to the search page. After I finish the search page I just need to add overlays to it and then move on to the playlist page.